--- a/test/parallel/taskCompare/elliot/empty-qthreads-chpl-like.chpl
+++ b/test/parallel/taskCompare/elliot/empty-qthreads-chpl-like.chpl
@@ -1,0 +1,18 @@
+use Time;
+
+config const numTrials = 100;
+config const printTimings = false;
+
+extern proc qtChplLikeTaskSpawn(trials, numTasks);
+
+proc main() {
+  var t: Timer;
+
+  t.start();
+  qtChplLikeTaskSpawn(numTrials, here.maxTaskPar);
+  t.stop();
+
+  if printTimings {
+    writeln("Elapsed time: ", t.elapsed());
+  }
+}