--- a/test/parallel/taskCompare/elliot/empty-qthreads-chpl-like.compopts
+++ b/test/parallel/taskCompare/elliot/empty-qthreads-chpl-like.compopts
@@ -1,0 +1,1 @@
+qtTaskSpawn.h