--- a/test/parallel/taskCompare/elliot/empty-qthreads-chpl-like.perfcompopts
+++ b/test/parallel/taskCompare/elliot/empty-qthreads-chpl-like.perfcompopts
@@ -1,0 +1,1 @@
+qtTaskSpawn.h