--- a/test/parallel/taskCompare/elliot/empty-qthreads-chpl-like.perfkeys
+++ b/test/parallel/taskCompare/elliot/empty-qthreads-chpl-like.perfkeys
@@ -1,0 +1,1 @@
+Elapsed time: