--- a/test/parallel/taskCompare/elliot/taskSpawn.graph
+++ b/test/parallel/taskCompare/elliot/taskSpawn.graph
@@ -1,5 +1,5 @@
-perfkeys: Elapsed time:, Elapsed time:, Elapsed time:, Elapsed time:
-graphkeys: forall, coforall, for+begin, omp parallel-for
-files: empty-forall.dat, empty-coforall.dat, empty-for+begin.dat, empty-omp-parallel-for.dat
+perfkeys: Elapsed time:, Elapsed time:, Elapsed time:, Elapsed time:, Elapsed time:
+graphkeys: forall, coforall, for+begin, omp parallel-for, chpl-like qthreads
+files: empty-forall.dat, empty-coforall.dat, empty-for+begin.dat, empty-omp-parallel-for.dat, empty-qthreads-chpl-like.dat
 graphtitle: Empty Task Spawn Timings (500,000 x maxTaskPar)
 ylabel: Time (seconds)