This is a raw qthreads version that mimics how Chapel spawns and waits for
tasks (qt_fork per task + an atomic counter)

Happily this version is competitive with the OpenMP variant on chapcs00: 3.2s
vs 3.1s. This indicates that qthreads itself doesn't have much overhead
compared to OpenMP, so we should be able to get our task spawning times to
match OpenMP with some work on our end. Note that the 3.2s requires using
spinwaiting. With condwait, the timing is more like 16s.